requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytest==6.1.1
pytest-django==3.10.0
pylint == 2.6.0